EP11: Fix host library version query
Look at release and modification level, not just the modification level.
Release and modification level are encoded into the one byte minor
field of a CK_VERSION. The high order 4 bits are the release number, the
low order 4 bits the modification level.

This allows host library version checks for release and modification levels.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>