CCA: Get master key verification patterns in token initialization
Obtain the master key verification patterns of the CCA master keys
(SYM, AES, and APKA) and check if all used APQNs are set up with the
same master keys (per master key type). If the expected master key
verification patterns are specified in the CCA token configuration file,
then also check if the master keys match the expected ones.

In case of mismatches, the token initialization will fail, and an
appropriate syslog message is issued.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>