ICA: Ensure to dlopen already loaded libica library
The ICA token links to the libica, but also load the libica dynamically
via dlopen() to obtain some optional symbols. Iterate over the already
loaded libraries and obtain the library handle of libica that way.
This ensures that the dynamically loaded libica is the exact same one
as the one linked with.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>