CCA: Introduce CCA token specific configuration file
To allow the user to specify the expected CCA master key verification
patterns, a new CCA specific configuration file is introduced.
For backward compatibility, the configuration file is optional.
The name of the configuration file can be specified via keyword
'confname' in opencryptoki.conf for a CCA token. This can either be
an absolute path, or a relative file name. For a relative file name,
the configuration file must exist in opencryptoki's configuration
directory, i.e. /etc/opencryptoki/.

The CCA token configuration file looks like this:

    version cca-0

    EXPECTED_MKVPS
    {
      SYM = "<SYM mkvp as 8 bytes hex string>"
      AES = "<AES mkvp as 8 bytes hex string>"
      APKA = "<APKA mkvp as 8 bytes hex string>"
    }

The specification of the EXPECTED_MKVPS structure is optional, also each
of the 3 master key verification patterns are optional. Each master key
verification pattern must be specified as an 16 character hex string
(8 bytes).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>