EP11: Parse EP11 token config file with generic config parser
The syntax of the EP11 token file stays the same, but for the following
keywords the 'key = value' and 'key = "value"' form is also accepted:
CPFILTER, PKEY_MODE, DIGEST_LIBICA.

Furthermore, end of line comments (using '#') are now accepted at any
places. Previously such comments were only accepted at the beginning of
a line.

Support for keyword 'LOGLEVEL <value>' is removed. This was accepted but
ignored since OpenCryptoki v3.3 (in 2015). It was issuing a syslog message
that LOGLEVEL is no longer supported, but ignored the setting since then.
Finally drop support for the 'LOGLEVEL' keyword completely.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>