EP11: Add config option to specify the expected WKVP
The expected wrapping key verification pattern can be specified
in the EP11 token config file as follows:

    EXPECTED_WKVP "<wkvp as 16 bytes hex string>"

The hex string must be quoted to not be parsed as a number. Numbers
can only be up to 64 bits (8 bytes), but the WKVP is 16 bytes.

The newer key-value-type specification is also supported, but the
hex string must also be quoted:

    EXPECTED_WKVP = "<wkvp as 16 bytes hex string>"

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>