EP11: Get wrapping key verification patterns in token initialization
Obtain the wrapping key verification patterns of the EP11 wrapping keys
and check if all APQNs configured for the token are set up with the
same wrapping key.

If an expected wrapping key verification pattern is specified in the
EP11 token configuration file, then also check if the wrapping keys
match the expected one.

In case of mismatches, the token initialization will fail, and an
appropriate syslog message is issued.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>