API: Trace OpenSSL error stack on return from API function
In case an OpenSSL function failed inside of the API function,
the OpenSSL error stack should show the reasons for the failure.

Please note that an error on the OpenSSL error stack does not necessarily
mean that an error has happened that would cause the API function to fail.
Some error situations are expected and internally handled, and do not lead
to a failure of the API function.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>