pkcshsm_mk_change: reencipher command
Implement the reencipher command of the pkcshsm_mk_change tool.
The reencipher command initiates a concurrent HSM master key change
operation. It gets the set of APQNs for which the master keys are to
be changed, as well as the master key verification patterns of the
new, to be set master key, for all master key types that are changed.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>