EP11: HSM MK change: Allow and check new WK if operation is active
The EP11 token checks the APQNs wrapping key verification pattern (WKVP) during
initialization and whenever a new APQN becomes online to be the expected
one. The expected WKVP is either specified in the EP11 token configuration file,
or is automatically determined during initialization. In any case, all APQNs
used by the token must have the same WKVP.

When an HSM master key change operation is active for the token, then the
WKVP may either be the one that was configured or determined, or it may already
be the WKVP of the new, to be set wrapping key specified in the master key change
operation, when the new WK has already been set/activated for an APQN.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>