EP11: HSM MK change: Select single APQN when operation is active
When an HSM master key change operation is active for a token, select
a single (online) APQN out of the APQNs configured for the token,
preferably one that already has the new WK set.

This simplifies the handling for the situation where the new WK is
activated on the APNQs by the HSM security officer.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>