EP11: HSM MK change: Re-encipher wrap blob if operation active
The EP11 token generates a random AES secure key (the so called wrap
blob) during token initialization that is used to wrap clear keys to
be imported.

When an HSM master change operation is active when the wrap blob is
generated, it is re-enciphered with the new WK right after creation.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>