EP11: HSM MK change: handle MK change event: finalize/cancel query
The pkcshsm_mk_change tool sends the EVENT_TYPE_MK_CHANGE_FINALIZE_QUERY
or EVENT_TYPE_MK_CHANGE_CANCEL_QUERY events to all tokens to check if
the tokens are able to finalize or cancel the an HSM master key change
operation.

For finalizing, the tokens check if the APQNs affected by the master key
change operation all have the new WK set/activated. It also checks if
the EP11 token config file specifies an expected WKVP, and if so, that
it has been updated to specify the WKVP of the new WK.

For canceling, the tokens check if the APQNs affected by the master key
change operation all still have the old WK set/activated. It also checks
if the EP11 token config file specifies an expected WKVP, and if so, that
it still specifies the WKVP of the old WK.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>