EP11: HSM MK change: retry logic if single APQN went offline
When an HSM master key change operation is active, the token operates on
a single APQN selected out of the set of APQNs configured for the token.

When the currently selected single APQN gets offline, the EP11 host library
calls fail. If the error is due to the single APQN being no longer available,
select a new online single APQN, out of the set of APQNs configured for the
token, and retry the library call.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>