common: Update SHM counters in object after loading
When a token object is loaded from disk, also update the SHM
counters in the object, so that it is not being reloaded in case
the counters in SHM are not zero, i.e. if the object has been
updated previously.

Also refactor object_mgr_restore_obj_withSize() while at it
to make it better readable.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>