common: Allow object_mgr_check_shm() with WRITE locked objs
For historical reasons object_mgr_check_shm() needed to be called
with an object that is READ locked. However, object_mgr_check_shm()
internally may have to re-lock the object for WRITE anyway.

Allow the caller to WRITE lock the object already, and pass the
lock type to object_mgr_check_shm(). That way, it may not have to
re-lock the object if it is already WRITE locked.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>