EP11: HSM MK change: Retry key operation for WK mismatch
If a crypto operation with a secure key fails with CKR_IBM_WKID_MISMATCH,
and an MK change is active, retry the operation with the re-enciphered
key blob from CKA_IBM_OPAQUE_REENC. If the retried operation succeeds,
this is an indication that the single APQN now has the new WK set.

If the single APQN has already the new WK set when the crypto operation
is attempted, then use the re-enciphered key blob from CKA_IBM_OPAQUE_REENC
right away.

Similar handling is done for crypto operations with the wrap blob.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>