EP11: HSM MK change: Re-encipher keyblob after set attributes
When the attributes of a key object are changed by means of C_SetAttributeValue,
and the changed attributes affect the secure key blob, then the secure key
blob may have to be re-enciphered afterwards if an HSM master key change
operation is active.

If the new WK is updated/set on the single APQN after the secure key blob
was updated, but before it was re-enciphered, then the operation is re-tried
with the (still unchanged) re-enciphered key blob from CKA_IBM_OPAQUE_REENC.
The resulting secure key blob is then enciphered with the new WK already.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>