EP11: HSM MK change: Retry logic for special case
When a crypto operation uses a token object, which was created by another
process that uses a single APQN that already has the new WK set, the key
blob of the token object is already enciphered with the new WK. In case
the current process runs with a single APQN that has not yet the new WK
set, the crypto operation fails with CKR_IBM_WKID_MISMATCH. Try to select
another single APQN which has the new WK set and retry the operation.
In case no APQN is available that has the new WK, wait until one becomes
available. This can only happen if the APQN with which the other process
created the token object, and already had the new WK activated, became
offline. It is expected that the other APQNs will also get the new WK
activated quickly.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>