EP11: HSM MK change: Re-encipher state blobs and handle state WK mismatch
For EP11 host library calls that maintain a state, it is required that
the state blob is re-enciphered after every state changing operation if
an HSM master key change operation is active.

If the new WK is updated/set on the single APQN after the state blob was
updated, but before it was re-enciphered, then the operation is re-tried
with the (still unchanged) re-enciphered state blob. The resulting state
blob is then enciphered with the new WK already.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>