EP11: HSM MK change: re-encipher newly created keys
When a HSM master key change operation is active, newly created secure
key blobs may have to be re-enciphered right after creation.

If the new WK is updated/set on the single APQN after the secure key was
created, but before it was re-enciphered, then the just created secure key
blob is thrown away, and the key creation operation is repeated.
The now created secure key blob is then enciphered with the new WK already.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>