common: Add support for C_SessionCancel function (PKCS#11 v3.0)
C_SessionCancel terminates active session based operations.

To identify which operation(s) should be terminated, the flags parameter
should be assigned the logical bitwise OR of one or more of the bit
flags defined in the CK_MECHANISM_INFO structure.

If no flags are set, the session state will not be modified and CKR_OK
will be returned.

If a flag is set for an operation that has not been initialized in the
session, the operation flag will be ignored and C_SessionCancel will
behave as if the operation flag was not set.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>