common: Fix C_Get/SetOperationState with multiple operations active
Although it is not stated explicitly in the PKCS#11 standard, but in a
session multiple operations may be active at a time. I.e. there can be a
digest operation being active and at the same time a sign operation.

This is used by the so called dual-function cryptographic functions like
C_DigestEncryptUpdate and friends. This require that the application
initialize a digest operation as well as an encrypt operation in the same
session.

However, there can only be one operations of the same type be active at
a time, i.e. there can not be two sign operations active in one session.

The current implementation in session_mgr_get_op_state() and
session_mgr_set_op_state() serializes/deserializes only one active
operation and fails in case of additional active operations.

Fix this by serializing/deserializing all active operations into/from the
state blob. The state data of additional active operations are appended
to the state blob.

Also add information about the openCryptoki version, the token model and
manufacturer to the state data. This allows to check if a user provided
state blob is compatible with the current session and token. Different
token models or token versions may have incompatible operation state data,
so do not try to apply an incompatible state data to a session, this would
produce unpredictable results, including segfaults.

Closes: https://github.com/opencryptoki/opencryptoki/issues/574

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>