common: Build without support for MD2 by default
MD2 is considered very weak, and has been disabled in many other
crypto libraries and applications in the meantime.

Build openCryptoki without support for MD2 by default. If really
needed, one can enable it with the '--enable-md2' configure option.

Only the soft token did support CKM_MD2, and CKM_MD2_HMAC[_GENERAL]
which is now built without support for MD2.

Note: MD5 is also weak, but it is still needed in openCrypoki for the
old (pre 3.12) token data format. The new data format is the default,
but there might still be installations using the old data format, so
we need to keep support for MD5 for the time being.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>