testcases: abfunc: Print BEGIN message even if entire test is skipped
Attribute bound keys are only supported by the EP11 token. When the
abfunc testcase is run against a non-EP11 token, the entire testcase
is skipped.

Print a "BEGIN xxx TESTCASE" message anyway, because otherwise the
output post-processor for the CI will not see this as a new testcase,
and report this skipped testcase under the testcase name from the
previously run testcase.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>