common: Helper function to iterate over active operations
Function session_mgr_iterate_session_ops() iterates over all active
sessions (if no specific session is specified) and calls a callback
function for each active operation in the session(s).

The callback function gets information about the session, the operation
type (digest, sign, verify, encrypt, decrypt), the operations key (if
any), and the context data and length.

For thread safety, the write lock of hsm_mk_change_rwlock should
be held be the caller. This ensures that the sessions and operations
are not modified concurrently.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>