CCA: Extract helper function to iterate over adapters/domains
Helper function to iterate over the CCA adapters and domains based on
the current setting of the CCA adapter and domain selection environment
variables CSU_DEFAULT_ADAPTER and CSU_DEFAULT_DOMAIN. With an environment
variable setting of 'DEV-ANY' and/or 'DOM-ANY' multiple adapters and/or
domains can be selected.

The callback function is called for each selected adapter/domain pair.
The respective adapter and domain has been allocated via CCA verb CSUACRA
when the callback function is called and is released via CCA verb CSUACRD
on return from the callback function. The callback function also gets the
APQN (cards and domain number) of the selected adapter and domain.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>