CCA: HSM MK change: handle MK change event: Initial query
The pkcshsm_mk_change tool sends the EVENT_TYPE_MK_CHANGE_INITIATE_QUERY
event to all tokens to check if the tokens are affected by a to be
initiated HSM master key change operation, and if so if the tokens can
fulfill it.

The tokens check if the set of APQNs affected by the master key change
operation is consistent with the tokens APQN configuration. All APQNs that
the token uses must be part of the master key change operation.
Also check if the affected APQNs all have the same new master key loaded.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>