CCA: HSM MK change: Allow and check new MKs if operation is active
The CCA token checks the APQNs master key verification patterns (MKVPs) during
initialization and whenever a new APQN becomes online to be the expected ones.
The expected MKVPs are either specified in the CCA token configuration file, or
are automatically determined during initialization. In any case, all APQNs used
by the token must have the same MKVPs.

When an HSM master key change operation is active for the token, then the
MKVPs may either be those that were configured or determined, or they may already
be the MKVPs of the new, to be set master keys specified in the master key change
operation, when the new MK has already been set/activated for an APQN.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>