CCA: HSM MK change: handle MK change event: finalize/cancel query
The pkcshsm_mk_change tool sends the EVENT_TYPE_MK_CHANGE_FINALIZE_QUERY
or EVENT_TYPE_MK_CHANGE_CANCEL_QUERY events to all tokens to check if
the tokens are able to finalize or cancel the HSM master key change
operation.

For finalizing, the tokens check if the APQNs affected by the master key
change operation all have the new MKs set/activated. It also checks if
the CCA token config file specifies any expected MKVPs, and if so, that
they have been updated to specify the MKVPs of the new MKs.

For canceling, the tokens check if the APQNs affected by the master key
change operation all still have the old MKs set/activated. They also check
if the CCA token config file specifies any expected MKVPs, and if so, that
they still specify the MKVPs of the old MKs.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>