config: Make confignode_foreach() macro more error save
If macro confignode_foreach() is used with 2nd parameter (c) being NULL, it
should not try to dereference it.

This can happen when an empty structure element is parsed and the structure
element's value is not checked for NULL before using confignode_foreach().
Since this is an empty structure, confignode_foreach() should simply do
nothing in this case.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>