common/soft/ICA/CCA/TPM: Pass template to AES and DES key gen function
Pass the template of the to be generated key object to the token specific
AES and DES key generation function. This allows the token to add additional
attributes into the template during key generation.

This is required for the CCA token to be able to add the CKA_IBM_OPAQUE_REENC
attribute to the key template, when a concurrent master key change operation
is active and the re-enciphered version of the generated key must be added,
too.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>