CCA: HSM MK change: re-encipher newly created or imported keys
When a HSM master key change operation is active, newly created or imported
secure keys blobs may have to be re-enciphered right after creation/import.

If the new MK is updated/set on the single APQN after the secure key was
created/imported, but before it was re-enciphered, then a single CCA APQN
(preferably with the new MK set) is selected, and the re-encipher is retried.

If the selected single APQN still has the old MK set, then the secure key
is re-enciphered from CURRENT-to-NEW. However, if the selected single APQN
has the new MK set, then the secure key is re-enciphered from OLD-to-CURRENT.
The CURRENT MK register of that APQN then has the NEW MK, and the OLD MK
register has the previous MK.

In case the single selected APQN still has the old MK set, it may happen
that the re-enciphering fails again if the new MK is updated/set on that
APQN after it was selected, but before the re-encipher attempt is made.
Re-select a single APQN (preferable with the new MK set) and retry the
re-enciphering in that case. We should now find an APQN with the new MK
set/activated and can then perform a OLD-to-CURRENT re-enciphering.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>