CCA: HSM MK change: Retry logic for special case
When a crypto operation uses a token object, which was created by another
process that uses an APQN that already has the new MK set, the key
blob of the token object is already enciphered with the new MK. In case
the current process runs with an APQN that has not yet the new MK(s)
set, the crypto operation fails with return code 8 reason code 48 "One or
more keys has a master key verification pattern that is not valid".
Try to select a single APQN which has the new MK(s) of the desired MK
type(s) set and retry the operation. In case no APQN is available that
has the new MK(s), wait until one becomes available. This can only happen
if the APQN with which the other process created the token object, and
already had the new MK(s) activated, became offline. It is expected that
the other APQNs will also get the new MK(s) activated quickly.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>