configure: Do not link all executables with libber and libldap
configure.ac should not add '-llber -lldap' to the global LIBS because
this will cause all executables to be linked with these libraries.
Only the tokens need libber (the ICSF token also libldap), so explicitly
add the lib(s) to the LDFLAGS of the tokens instead of adding both
globally.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>