EP11: Handle m_Get/m_SetAttributeValue with session bound blobs
Like other EP11 host library functions that operate on key blobs,
m_GetAttributeValue and m_SetAttributeValue may also return
CKR_SESSION_CLOSED when the session that the key blob is bound to
is not logged in into the adapter that is currently used for the
request.

For that the session must be passed through to the functions.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>