configure: Remove superfluous --with-xcryptolinz configure option
The --with-xcryptolinz configure option was used for specifying the CCA
library development files location, but the CCA token does not include
any of the CCA header files. It has all required CCA definitions in its
own header files. Also, the CCA library is loaded via dlopen, and is not
linked with the CCA token, so no CCA development files are required at all
to build the CCA token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>