p11sak: Allow to set different labels for public and private keys
By default, when generating an asymmetric key pair, the specified label
is appended by ":pub" and ".prv" for the public and private key object.

Allow the user to explicitly specify the label for public and private
keys objects. The separate character is a colon (':'). Anything before the
colon is the public key label, anything after it is the private key label.

To set the exact same label for private and public key objects, use
"pub-label:=". The equal character means hat the private label is the same
as the public label.

In case a colon or a equal character occurs within a label, it must be
escaped by a backslash ('\'). Note that the shell may interpret escape
characters as well, so better quote the label specification.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>