EP11: Correct precondition checks for protected key usage and creation
Using an existing and valid protected key attached to a key object is
not dependent on the session state. Only if the protected key must be
(re-)created (secure key to protected key transformation), a proper session
state is required.

PKCS#11-UG states:
"A session can be a read/write (R/W) session or a read-only (R/O) session.
Read/write and read-only refer to the access to token objects, not to session
objects. In both session types, an application can create, read, write and
destroy session objects, and read token objects. However, only in a read/write
session can an application create, modify, and destroy token objects."

Allow the transformation of a key object into a protected key for session
objects also in a read-only session. As per above, session objects can be
updated in read-only session.

Furthermore PKCS#11-UG states for SO-RW sessions (CKS_RW_SO_FUNCTIONS):
"The Security Officer has been authenticated to the token. The application
has read/write access only to public objects on the token, not to private
objects."

Thus, in a SO read/write session, private key objects can not be transformed
into a protected key.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>