ci: Match APQNs that contain hex digits, not only decimal
When running in the CI (make ci-installcheck), slot 46 is only defined
when there a CEX7P APQNs available. Checking for decimal digits only will
cause it to not create that slot when all APQNs have hex digits in its
addresses (which is the case on almost all systems in the CI).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>