policy: Fix update of mechanism info for AES-XTS mechanisms
AES-XTS keys use 2 keys, and thus the AES-XTS mechanisms report the minimum
and maximum key size in the mechanism info as 32 - 64 bytes (256 - 512 bits).
However, the strength of such an AES-XTS key is just the half of its size,
i.e. a 256 bit AES-XTS key has a strength of 128 bits only. Likewise a 512 bit
AES-XTS key has a strength of 256 bits.

While the checking of the key strength when using an AES-XTS key is done
correctly, the supported minimum key size in the mechanism list is not updated
according to the minimum strength policy setting. So, if a minimum strength of
256 bits is configured, then the minimum key size of the AES-XTS mechanisms must
be adjusted to 64 bytes (512 bits).

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>