ICA: Calculate private exponent when generating RSA keys
The ICA token uses libica to generate RSA keys in CRT format. That way, the
public key components n and e, as well as the private CTR components p, q,
dp, dq, and qInverse are available, but not the private exponent d.

Calculate the private exponent d from the modulus n, the public exponent e,
and the private key components p and q as follows:

ϕ(n) = (p − 1) * (q − 1) = n − p − q + 1. Then d = e ^−1 mod ϕ(n)

That way, an RSA key generated by the ICA token has all private key components,
the mod-expo components as well as the CRT components.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>